Initialize a keyATM model

This function creates a list of word indexes W, topic indicators Z, seed indicators X, a vocabulary list vocab from files.

keyATM_model(files = NULL, dict = NULL, text_df = NULL,
  text_dfm = NULL, mode = "basic", extra_k = 1,
  covariates_data = NULL, covariates_formula = NULL,
  num_states = NULL, timestamps = NULL, encoding = "UTF-8",
  lowercase = TRUE, remove_numbers = TRUE, remove_punct = TRUE,
  remove_symbols = TRUE, remove_separators = TRUE,
  remove_twitter = FALSE, remove_hyphens = FALSE, remove_url = TRUE,
  stem_language = NULL, stopwords = NULL, alpha = 50/(length(dict) +
  extra_k), beta = 0.01, beta_s = 0.1, gamma_1 = 1, gamma_2 = 1,
  options = list())

Arguments

files
names of each file to read (or a quanteda corpus object)

dict
a quanteda dictionary or named list of character vectors

text_df
directly passes a text in a data.frame

mode
"basic", "cov", "tot", or "hmm"

extra_k
number of unseeded topics in addition to the topics seeded by dict

covariates_data
a data.frame or a tibble that is a covariate matrix. Columns are covariates.

covariates_formula
formula for the covariates, for example, ~. uses all variables

timestamps
time data

encoding
File encoding (Default: whatever quanteda guesses)

lowercase
whether to transform each token to lowercase letters

remove_numbers
whether to remove numbers

remove_punct
whether to remove punctuation

remove_symbols
whether to remove non-alphanumerical symbols

remove_separators
whether to remove tabs, spaces, newlines, etc.

remove_twitter
whether to remove Twitter detritus

remove_hyphens
whether to split hyphenated words

remove_url
whether to remove URLs

stem_language
if not NULL, the language to use for stemming

stopwords
if not NULL, a character vector of words to remove, e.g. quanteda::stopwords("english")

alpha
Starting value for all the model's topic proportion hyperparameters. Default: 50 / number of topics)

beta
Hyperparameter for estimated word probabilities. Default: 0.01

beta_s
Hyperparameter for seeded word probabilities. Default: 0.1

gamma_1
First Beta hyperparameter for probability of being drawn from a seeded topic. Default: 1.0

gamma_2
Second Beta hyperparameter for probability of being drawn from a seeded topic. Default: 1.0

dtm
Document-Term matrix from quanteda package

num_state
numer of state in HMM model

Value

A list containing

W

a list of vectors of word indexes

Z

a list of vectors of topic indicators isomorphic to W

C

a covariate matrix is there is an input

X

a list of vectors of seed indicators (0/1) isomorphic to W

vocab

a vector of vocabulary items

files

a vector of document filenames

dict

a tokenized version of the dictionary

seeds

a list of words for the seed words in dict, named by dictionary category

extra_k

how many extra non-seeded topics are required

alpha

a vector of topic proportion hyperparameters. If you use the model with covariates, it is not used.

alpha_iter

a list to store topic proportion hyperparameters

Lambda_iter

a list to store coefficients of the covariates

S_iter

a list to store states sampled in HMM

tot_beta

a list to store sampled beta parameters in topic-over-time

sampling_info

information related to sampling

model_fit

a list to store perplexity and log-likelihood

gamma1

First prior probability parameter for X (currently the same for all topics)

gamma2

Second prior probability parameter for X (currently the same for all topics)

beta

prior parameter for the non-seeded word generation probabilities

beta_s

prior parameter for the seeded word generation probabilities

use_cov

boolean, whether or not use the covariate

num_states

number of states in HMM

timestamps

time stamp for topic-over-time model

visualize_keywords

ggplot2 object

call

details of the function call

.

Details

Z represents which topic a word topken was generated by. It is initialized randomly from length(dict) + extra_k topics, except when dict assigns a word to a particular category, in which case this topic indicator is assigned. For example, if a word appears in the second category of dict it will always be initialized as 1.

X is 1 when a word is generated from one of the topics from the dictionary ('seeded') and 0 when it was generated from one of the extra_k normal topics. X is initialized randomly except when the word is contained in dict.

Note that all indicators are zero-based for ease of later processing in C++, so mod$vocab[mod$W + 1] recovers the (tokenized) words of the i-th document from model mod.

If alpha is a scalar this is the value given to all elements of alpha. If it is a vector of the correct length those values are used as the starting alphas.