Predict topic proportions for the covariate keyATM

Predict topic proportions for the covariate keyATM

# S3 method for keyATM_output
predict(
  object,
  newdata,
  transform = FALSE,
  burn_in = NULL,
  parallel = TRUE,
  posterior_mean = TRUE,
  ci = 0.9,
  method = c("hdi", "eti"),
  point = c("mean", "median"),
  label = NULL,
  raw_values = FALSE,
  ...
)

Arguments

object

the keyATM_output object for the covariate model.

newdata

New observations which should be predicted.

transform

Transorm and standardize the newdata with the same formula and option as model_settings used in keyATM().

burn_in

integer. Burn-in period. If not specified, it is the half of samples. Default is NULL.

parallel

logical. If TRUE, parallelization for speeding up. Default is TRUE. Please plan() before use this function.

posterior_mean

logical. If TRUE, the quantity of interest to estimate is the posterior mean. Default is TRUE.

ci

value of the credible interval (between 0 and 1) to be estimated. Default is 0.9 (90%).

method

method for computing the credible interval. The Highest Density Interval (hdi, default) or Equal-tailed Interval (eti).

point

method for computing the point estimate. mean (default) or median.

label

a character. Add a label column to the output. The default is NULL (do not add it).

raw_values

a logical. Returns raw values. The default is FALSE.

...

additional arguments not used.